Install script: don't download composer if already downloaded

+++ install.php
@@ -1,11 +1,13 @@
 <?php
 
-echo "Downloading Composer installer..." . PHP_EOL;
-@mkdir('build');
-file_put_contents("build/composer_installer.php", file_get_contents('http://getcomposer.org/installer'));
+if (!file_exists('build/composer.phar')) {
+    echo "Downloading Composer installer..." . PHP_EOL;
+    @mkdir('build');
+    file_put_contents("build/composer_installer.php", file_get_contents('http://getcomposer.org/installer'));
 
-echo "Installing composer.phar" . PHP_EOL;
-echo shell_exec("php build/composer_installer.php --install-dir build");
+    echo "Installing composer.phar" . PHP_EOL;
+    echo shell_exec("php build/composer_installer.php --install-dir build");
+}
 
 echo "Installing dependencies" . PHP_EOL;
 echo shell_exec("php build/composer.phar install --dev");
